refactor: cleanup const and types
