+++ packages/swapper/src/swappers/ChainflipSwapper/ChainflipSwapper.ts
@@ -6,7 +6,6 @@ import { CHAINFLIP_SUPPORTED_CHAIN_IDS } from './constants'
 
 export const chainflipSwapper: Swapper = {
   filterAssetIdsBySellable: (assets: Asset[]): Promise<AssetId[]> => {
-    // TODO: Return only Chainflip supported coins
     return Promise.resolve(
       assets
         .filter(asset => CHAINFLIP_SUPPORTED_CHAIN_IDS.sell.includes(asset.chainId))
@@ -15,7 +14,6 @@ export const chainflipSwapper: Swapper = {
   },
 
   filterBuyAssetsBySellAssetId: (input: BuyAssetBySellIdInput): Promise<AssetId[]> => {
-    // TODO: Return only Chainflip supported coins
     return Promise.resolve(
       input.assets
         .filter(asset => CHAINFLIP_SUPPORTED_CHAIN_IDS.buy.includes(asset.chainId))
